Initial Commit
Amazon, the world's largest online retailer, offers an extensive range of products across multiple categories such as electronics, fashion, home goods, and more. In today's competitive business landscape, it is essential to gain insights into which products are performing well and why, to stay ahead of the competition and ensure customer satisfaction. This project aims to analyze Amazon's sales data to identify the top-performing categories and products, as well as the factors that influence their sales. Our analysis can be useful for Amazon and other businesses to optimize their product offerings, marketing strategies, and inventory management, leading to improved sales and customer satisfaction.